+++ packages/core/src/objects/HexString.ts
@@ -1,12 +1,15 @@
 import { Convert } from "pvtsutils";
-import { BadCharError } from "../BadCharError";
+import { BadCharError } from "../errors";
 import { CharSet } from "../CharSet";
 import type { ViewReader } from "../ViewReader";
 import type { ViewWriter } from "../ViewWriter";
 import { TextEncoder } from "./TextEncoder";
 import { PDFTextString } from "./TextString";
+import { ObjectTypeEnum } from "./internal";
 
 export class PDFHexString extends PDFTextString {
+
+  public static readonly NAME = ObjectTypeEnum.HexString;
 
   protected onWritePDF(writer: ViewWriter): void {
     const encText = TextEncoder.to(this.text);