+++ packages/core/src/objects/Indirect.ts
@@ -1,4 +1,5 @@
-import { BadCharError } from "../BadCharError";
+import { BadCharError } from "../errors";
+import { CharSet } from "../CharSet";
 import type { ViewReader } from "../ViewReader";
 import { IPDFIndirect, PDFObject } from "./Object";
 
@@ -14,6 +15,20 @@ export abstract class PDFIndirect extends PDFObject implements IPDFIndirect {
   }
 
   protected onFromPDF(reader: ViewReader): void {
+    while (true) {
+      // skip white spaces
+      reader.findIndex(c => !CharSet.whiteSpaceChars.includes(c));
+      if (reader.isEOF) {
+        break;
+      }
+      // skip comment
+      if (reader.view[reader.position] === CharSet.percentChar) {
+        PDFComment.fromPDF(reader);
+        continue; // continue white spaces and comment skipping if comment found
+      }
+      break;
+    }
+
     const objectNumber = PDFNumeric.fromPDF(reader);
     PDFNumeric.assertPositiveInteger(objectNumber);
 
@@ -48,3 +63,5 @@ export abstract class PDFIndirect extends PDFObject implements IPDFIndirect {
 }
 
 import { PDFNumeric } from "./Numeric";
+import { PDFComment } from "./Comment";
+