+++ packages/core/src/objects/LiteralString.ts
@@ -1,8 +1,8 @@
 import { Convert } from "pvtsutils";
+import { BadCharError, ParsingError } from "../errors";
 import type { ViewReader } from "../ViewReader";
 import type { ViewWriter } from "../ViewWriter";
-import { BadCharError } from "../BadCharError";
-import { ParsingError } from "../ParsingError";
+import { ObjectTypeEnum } from "./internal";
 import { PDFTextString } from "./TextString";
 import { TextEncoder } from "./TextEncoder";
 
@@ -10,6 +10,8 @@ const leftParenthesisChar = 0x28;
 const rightParenthesisChar = 0x29;
 
 export class PDFLiteralString extends PDFTextString {
+
+  public static readonly NAME = ObjectTypeEnum.LiteralString;
 
   protected onWritePDF(writer: ViewWriter): void {