+++ packages/core/src/objects/Stream.ts
@@ -4,15 +4,13 @@ import type { ViewReader } from "../ViewReader";
 import type { ViewWriter } from "../ViewWriter";
 import type { EncryptionObject } from "./EncryptionObject";
 
-import { BadCharError } from "../BadCharError";
-import { ParsingError } from "../ParsingError";
+import { BadCharError, ParsingError } from "../errors";
+import { ObjectTypeEnum, typeOf } from "./internal";
 import { PDFDictionary } from "./Dictionary";
-import { PDFDictionaryField } from "./DictionaryFieldDecorator";
-
+import { PDFDictionaryField } from "./decorators/field";
 import { PDFName } from "./Name";
-import { PDFNull } from "./Null";
 import { PDFNumeric } from "./Numeric";
-import { PDFArray } from "./Array";
+import type { PDFArray } from "./Array";
 
 const streamChars = new Uint8Array([0x73, 0x74, 0x72, 0x65, 0x61, 0x6d]);
 const endStreamChars = new Uint8Array([0x65, 0x6e, 0x64, 0x73, 0x74, 0x72, 0x65, 0x61, 0x6d]);
@@ -20,15 +18,15 @@ const endStreamChars = new Uint8Array([0x65, 0x6e, 0x64, 0x73, 0x74, 0x72, 0x65,
 //#region Filed handlers
 
 function getDecodeParamsHandler(this: PDFStream, obj: PDFObjectTypes): Array<PDFDictionary | null> {
-  if (obj instanceof PDFDictionary) {
+  if (typeOf(obj, ObjectTypeEnum.Dictionary)) {
     return [obj];
-  } else if (obj instanceof PDFArray) {
+  } else if (typeOf(obj, ObjectTypeEnum.Array)) {
     const res: Array<PDFDictionary | null> = [];
 
     for (const item of obj.items) {
-      if (item instanceof PDFDictionary) {
+      if (typeOf(item, ObjectTypeEnum.Dictionary)) {
         res.push(item);
-      } else if (item instanceof PDFNull) {
+      } else if (typeOf(item, ObjectTypeEnum.Null)) {
         res.push(null);
       } else {
         throw new TypeError("Unsupported type of DecodeParms subitem in the PDF Stream");
@@ -44,6 +42,8 @@ function getDecodeParamsHandler(this: PDFStream, obj: PDFObjectTypes): Array<PDF
 //#endregion
 
 export class PDFStream extends PDFDictionary implements EncryptionObject {
+
+  public static override readonly NAME = ObjectTypeEnum.Stream;
 
   protected static skipEndOfLine(reader: ViewReader): void {
     const view = reader.view.subarray(reader.position);
@@ -94,7 +94,8 @@ export class PDFStream extends PDFDictionary implements EncryptionObject {
       if (filters.length === 1) {
         stream.filter = new PDFName(filters[0].name);
       } else {
-        stream.filter = new PDFArray(...filters.map(o => new PDFName(o.name)));
+        const ArrayConstructor = PDFObjectReader.get(ObjectTypeEnum.Array);
+        stream.filter = new ArrayConstructor(...filters.map(o => new PDFName(o.name)));
       }
     }
 
@@ -211,11 +212,11 @@ export class PDFStream extends PDFDictionary implements EncryptionObject {
 
     if (this.filter) {
       let filterNames: string[] = [];
-      if (this.filter instanceof PDFName) {
+      if (typeOf(this.filter, ObjectTypeEnum.Name)) {
         filterNames.push(this.filter.text);
-      } else if (this.filter instanceof PDFArray) {
+      } else if (typeOf(this.filter, ObjectTypeEnum.Array)) {
         filterNames = this.filter.items.map(o => {
-          if (o instanceof PDFName) {
+          if (typeOf(o, ObjectTypeEnum.Name)) {
             return o.text;
           }
           throw new TypeError("Unsupported type of Filter in the PDF Stream");
@@ -310,7 +311,7 @@ export class PDFStream extends PDFDictionary implements EncryptionObject {
     if (!this.encrypted) {
       if (this.documentUpdate?.document.encryptHandler && !(this.has("Type") && this.get("Type", PDFName).text === "XRef")) {
         // The cross-reference stream shall not be encrypted and strings appearing in the cross-reference
-        // stream dictionary shall not be encrypted. It shall not have a Filter entry that specifies a Crypt filter 
+        // stream dictionary shall not be encrypted. It shall not have a Filter entry that specifies a Crypt filter
         const encryptedText = await this.encryptAsync();
         this.stream = BufferSourceConverter.toUint8Array(encryptedText);
         this.length.value = this.stream.length;
@@ -340,7 +341,7 @@ export class PDFStream extends PDFDictionary implements EncryptionObject {
                 let ref = "";
                 if (this.isIndirect()) {
                   const indirect = this.getIndirect();
-                  ref = ` (${indirect.id} ${indirect.generation} R)`;
+                  ref = ` (R ${indirect.id} ${indirect.generation})`;
                 }
                 throw new Error(`Cannot decrypt PDF stream${ref}. ${e.message}`);
               }
@@ -424,13 +425,13 @@ export class PDFStream extends PDFDictionary implements EncryptionObject {
   /**
    * Adds the name of the filter to the Field entry
    * @param name The name of the filter
-   * @returns 
+   * @returns
    */
   public addFilter(name: string): void {
-    if (this.filter instanceof PDFArray) {
+    if (typeOf(this.filter, ObjectTypeEnum.Array)) {
       // Array
       for (const filter of this.filter) {
-        if (filter instanceof PDFName && filter.text === name) {
+        if (typeOf(filter, ObjectTypeEnum.Name) && filter.text === name) {
           return;
         }
       }
@@ -468,6 +469,7 @@ export class PDFStream extends PDFDictionary implements EncryptionObject {
 
 }
 
-import { PDFObjectReader, PDFObjectTypes } from "./ObjectReader";
+import { PDFObjectReader } from "./ObjectReader";
+import { PDFObjectTypes } from "./ObjectTypes";
 import { Filter, FilterFactory } from "../filters";
 import { PDFIndirectObject } from "./IndirectObject";