+++ packages/core/src/structure/Document.spec.ts
@@ -2,7 +2,8 @@ import * as fs from "fs";
 import * as path from "path";
 import { PDFArray, PDFDictionary, PDFNumeric, PDFStream } from "../objects";
 import { PDFTextString } from "../objects/TextString";
-import { PDFDocument, XrefStructure } from "./Document";
+import { PDFDocument } from "./Document";
+import { XrefStructure } from "./XrefStructure";
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const pkijs = require("pkijs");
 import { Crypto } from "@peculiar/webcrypto";
@@ -12,8 +13,6 @@ import { BufferSourceConverter, Convert } from "pvtsutils";
 import * as assert from "assert";
 import { PDFDocumentObjectTypes } from "./DocumentObject";
 import { CrossReferenceStream } from "./CrossReferenceStream";
-const fetch = require("node-fetch") as typeof globalThis.fetch;
-
 
 const filesDir = path.join(__dirname, "..", "..", "..", "..", "files");